pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = "creating a rest API"
authors = ["yuenfu001 <liuyuenfu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
djangorestframework = "^3.14.0"
black = "^23.9.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
